services/filter_service/requirements.txt
fasttext
confluent-kafka
psycopg2-binary